add impersonation to user, and metadata to user and org (#2)
The User now has an impersonator_user_id field. This is the UUID of the other user that is impersonating this user. It will be None if impersonation is not happening (which it normally isn't).

Metadata is private, user-specific metadata which can be set via the regular update_metadata endpoint. The user will never see or use this metadata, it's just for the PropelAuth customer.

The Org objects have org_metadata. Much like user metadata, this is a private field.

The metadata objects are simple hashes with a string key and any JSON object for the value.